+++ R/atlas_species.R
@@ -87,7 +87,7 @@ atlas_species <- function(request = NULL,
   class(custom_call) <- "data_request"
   
   # check for caching
-  caching <- getOption("galah_config")$caching
+  caching <- getOption("galah_config")$package$caching
   cache_file <- cache_filename("species", unlist(custom_call))
   if (caching && file.exists(cache_file) && !refresh_cache) {
     return(read_cache_file(cache_file))
@@ -138,22 +138,22 @@ atlas_species_internal <- function(request,
     build_query(identify, filter, geolocate, profile = profile),
     emailNotify = email_notify(),
     sourceTypeId = 2004,
-    reasonTypeId = getOption("galah_config")$download_reason_id,
+    reasonTypeId = getOption("galah_config")$user$download_reason_id,
     email = user_email(), 
     facets = species_facets(),
     lookup = "true"
   )
   
   tmp <- tempfile()
-  url <- atlas_url("records_species")
-  result <- atlas_download(url, params = query, cache_file = tmp)
+  url <- url_lookup("records_species")
+  result <- url_download(url, params = query, cache_file = tmp, ext = "csv")
   
   if(is.null(result)){
     system_down_message("atlas_species")
     return(tibble())
   }else{
   
-    if(getOption("galah_config")$atlas == "Australia"){
+    if(getOption("galah_config")$atlas$region == "Australia"){
       # overwrite file with fixed names
       names(result) <- rename_columns(names(result), type = "checklist")
       result <- result[, wanted_columns("checklist")]