+++ R/url_lookup.R
@@ -1,19 +1,19 @@
-#' Internal function to build APIs
-#'
-#' This function does a couple of things. Trivially, it looks up 
-#' base URLs from an internal tibble (`node_config`), itself imported from 
-#' a csv (`./data-raw/node_config.csv`). Entries in that tibble can contain
-#' `glue`-like syntax for infilling new data, passed using `...`
-#'
+# Internal function to build APIs
+#
+# This function does a couple of things. Trivially, it looks up 
+# base URLs from an internal tibble (`node_config`), itself imported from 
+# a csv (`./data-raw/node_config.csv`). Entries in that tibble can contain
+# `glue`-like syntax for infilling new data, passed using `...`
+#
 #' @importFrom glue glue_data
 #' @keywords internal
 
-atlas_url <- function(api_name, ..., quiet = FALSE, error_call = caller_env()){
+url_lookup <- function(api_name, ..., quiet = FALSE, error_call = caller_env()){
   
   dots <- list(...)
   
   # run a check that a row in `node_config` matches your specification
-  current_atlas <- getOption("galah_config")$atlas
+  current_atlas <- getOption("galah_config")$atlas$region
   url_lookup <- node_config$api_name == api_name &
                 node_config$atlas == current_atlas